+++ core-ng/src/main/java/core/framework/internal/web/websocket/WebSocketCloseCodes.java
@@ -6,6 +6,7 @@ package core.framework.internal.web.websocket;
 public final class WebSocketCloseCodes {
     // refer to https://developer.mozilla.org/en-US/docs/Web/API/CloseEvent/code
     public static final int NORMAL_CLOSURE = 1000;
+    public static final int ABNORMAL_CLOSURE = 1006;
     public static final int POLICY_VIOLATION = 1008;
     public static final int INTERNAL_ERROR = 1011;
     public static final int SERVICE_RESTART = 1012;